+++ angular-workspace/components/icon/icon.stories.ts
@@ -4,7 +4,8 @@ import { IconArgs, iconMeta, iconStories } from "dso-toolkit";
 import icons from "dso-toolkit/storybook-assets/icons.json";
 
 import { DsoIcon } from "../../projects/component-library/src/public-api";
-import { templateContainer } from "../../templates";
+
+import { iconTemplate } from "./icon.angular-template";
 
 const meta: Meta<IconArgs> = {
   ...iconMeta({ readme }),
@@ -19,10 +20,7 @@ const meta: Meta<IconArgs> = {
 export default meta;
 
 const { Default } = iconStories({
-  templateContainer,
-  storyTemplates: (templates) => {
-    const { iconTemplate } = templates;
-
+  storyTemplates: () => {
     return {
       iconTemplate,
     };