+++ storybook/src/components/info/info.core-template.ts
@@ -1,5 +1,6 @@
 import { Info } from "dso-toolkit";
 import { TemplateResult, html } from "lit-html";
+import { ifDefined } from "lit-html/directives/if-defined.js";
 
 import { ComponentImplementation } from "../../templates";
 
@@ -7,7 +8,9 @@ export const coreInfo: ComponentImplementation<Info<TemplateResult>> = {
   component: "info",
   implementation: "core",
   template: () =>
-    function infoTemplate({ fixed, active, content, dsoClose }) {
-      return html`<dso-info ?fixed=${fixed} ?active=${active} @dsoClose=${dsoClose}>${content}</dso-info>`;
+    function infoTemplate({ fixed, active, content, dsoClose, id }) {
+      return html`<dso-info ?fixed=${fixed} ?active=${active} @dsoClose=${dsoClose} id=${ifDefined(id)}
+        >${content}</dso-info
+      >`;
     },
 };