+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Changed
 * Document Component: `markFn` markeert onterecht tekst in tooltip ([#3865](https://github.com/dso-toolkit/dso-toolkit/issues/3865))
 
+### Fixed
+* Cards: CardClick event wordt niet afgevuurd wanneer NVDA aan staat ([#3860](https://github.com/dso-toolkit/dso-toolkit/issues/3860))
+
 ### Removed
 * **BREAKING** Info: Remove HTML/CSS implementatie ([#3679](https://github.com/dso-toolkit/dso-toolkit/issues/3679))